backend/requirements.txt
# MediVision AI Backend - Requirements
# Production-optimized dependencies

# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Async Database
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
psycopg2-binary>=2.9.10
alembic>=1.14.0
aiosqlite>=0.20.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.19

# Redis & Background Tasks
redis>=5.2.0
celery>=5.4.0
flower>=2.0.1

# File Handling
python-magic>=0.4.27
aiofiles>=24.1.0
boto3>=1.35.0

# Medical Imaging
pydicom>=3.0.0
SimpleITK>=2.4.0
nibabel>=5.3.0

# Image Processing
numpy>=2.2.0
scipy>=1.15.0
Pillow>=11.0.0
opencv-python-headless>=4.10.0
scikit-image>=0.25.0

# Deep Learning
torch>=2.5.0
torchvision>=0.20.0
torchaudio>=2.5.0
transformers>=4.47.0
timm>=1.0.12
einops>=0.8.0

# Speech & Audio
openai-whisper>=20240930
librosa>=0.10.2
soundfile>=0.12.1
noisereduce>=3.0.3

# LLM & Fine-tuning
peft>=0.14.0
bitsandbytes>=0.45.0
accelerate>=1.2.0
datasets>=3.2.0

# Utilities
httpx>=0.28.0
python-dotenv>=1.0.1
loguru>=0.7.3
tqdm>=4.67.0
pyyaml>=6.0.2
orjson>=3.10.0

# PDF Generation
reportlab>=4.2.0
PyPDF2>=3.0.1

# Testing
pytest>=8.3.0
pytest-asyncio>=0.25.0
pytest-cov>=6.0.0

# Development & Linting
ruff>=0.8.0
mypy>=1.14.0
